+++ modulo3/intro-react/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "primeiro-react",
+  "name": "intro-react",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
