--- modulo3/intro-react/primeiro-react/src/App.css
@@ -1,35 +1,0 @@
-.App {
-    text-align: center;
-    background-color: #282c34;
-    min-height: 100vh;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    font-size: calc(10px + 2vmin);
-    color: white;
-}
-
-.App-perfil {
-    height: 40vmin;
-    pointer-events: none;
-    border-radius: 50%;
-}
-
-.App-link {
-    color: #61dafb;
-}
-
-.App-botao {
-    background-color: #61dafb;
-    color: #282c34;
-    padding: 10px 20px;
-    border-radius: 5px;
-    cursor: pointer;
-    opacity: 1;
-    transition: 0.3s;
-}
-
-.App-botao:hover {
-    opacity: 0.7;
-}
