+++ modulo3/intro-react/src/App.js
@@ -1,0 +1,12 @@
+import React from "react";
+import Home from "./Pages/Home";
+
+function App() {
+    return (
+        <div>
+            <Home />
+        </div>
+    );
+}
+
+export default App;
