+++ modulo3/intro-react/src/Pages/Home.css
@@ -1,0 +1,33 @@
+* {
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+}
+
+body {
+    background-color: #262837;
+    color: #ffffff;
+    font-family: Arial, Helvetica, sans-serif;
+}
+
+img {
+    width: 400px;
+}
+
+.container {
+    display: grid;
+    place-items: center;
+    height: 100vh;
+}
+
+.botao {
+    display: inline-block;
+    background-color: aqua;
+    padding: 10px 30px;
+    font-size: 1rem;
+    font-weight: bold;
+    text-transform: uppercase;
+    border: none;
+    border-radius: 5px;
+    cursor: pointer;
+}
